feat: change to f32 for performance

+++ src/diffusions/cir.rs
@@ -5,22 +5,22 @@ use crate::{
 use ndarray::Array1;
 
 pub fn cir(
-  theta: f64,
-  mu: f64,
-  sigma: f64,
+  theta: f32,
+  mu: f32,
+  sigma: f32,
   n: usize,
-  x0: Option<f64>,
-  t: Option<f64>,
+  x0: Option<f32>,
+  t: Option<f32>,
   use_sym: Option<bool>,
-) -> Vec<f64> {
+) -> Vec<f32> {
   if 2.0 * theta * mu < sigma.powi(2) {
     panic!("2 * theta * mu < sigma^2")
   }
 
   let gn = gn::gn(n - 1, Some(t.unwrap_or(1.0)));
-  let dt = t.unwrap_or(1.0) / n as f64;
+  let dt = t.unwrap_or(1.0) / n as f32;
 
-  let mut cir = Array1::<f64>::zeros(n);
+  let mut cir = Array1::<f32>::zeros(n);
   cir[0] = x0.unwrap_or(0.0);
 
   for i in 1..n {
@@ -36,15 +36,15 @@ pub fn cir(
 
 #[allow(clippy::too_many_arguments)]
 pub fn fcir(
-  hurst: f64,
-  theta: f64,
-  mu: f64,
-  sigma: f64,
+  hurst: f32,
+  theta: f32,
+  mu: f32,
+  sigma: f32,
   n: usize,
-  x0: Option<f64>,
-  t: Option<f64>,
+  x0: Option<f32>,
+  t: Option<f32>,
   use_sym: Option<bool>,
-) -> Vec<f64> {
+) -> Vec<f32> {
   if !(0.0..1.0).contains(&hurst) {
     panic!("Hurst parameter must be in (0, 1)")
   }
@@ -54,9 +54,9 @@ pub fn fcir(
   }
 
   let fgn = FgnFft::new(hurst, n - 1, t, None).sample();
-  let dt = t.unwrap_or(1.0) / n as f64;
+  let dt = t.unwrap_or(1.0) / n as f32;
 
-  let mut fcir = Array1::<f64>::zeros(n);
+  let mut fcir = Array1::<f32>::zeros(n);
   fcir[0] = x0.unwrap_or(0.0);
 
   for i in 1..n {
