+++ src/models/heston.rs
@@ -4,22 +4,22 @@ use crate::prelude::correlated::correlated_bms;
 
 #[allow(clippy::too_many_arguments)]
 pub fn heston(
-  mu: f64,
-  kappa: f64,
-  theta: f64,
-  eta: f64,
-  rho: f64,
+  mu: f32,
+  kappa: f32,
+  theta: f32,
+  eta: f32,
+  rho: f32,
   n: usize,
-  s0: Option<f64>,
-  v0: Option<f64>,
-  t: Option<f64>,
+  s0: Option<f32>,
+  v0: Option<f32>,
+  t: Option<f32>,
   use_sym: Option<bool>,
-) -> [Vec<f64>; 2] {
+) -> [Vec<f32>; 2] {
   let correlated_bms = correlated_bms(rho, n, t);
-  let dt = t.unwrap_or(1.0) / n as f64;
+  let dt = t.unwrap_or(1.0) / n as f32;
 
-  let mut s = Array1::<f64>::zeros(n);
-  let mut v = Array1::<f64>::zeros(n);
+  let mut s = Array1::<f32>::zeros(n);
+  let mut v = Array1::<f32>::zeros(n);
 
   s[0] = s0.unwrap_or(0.0);
   v[0] = v0.unwrap_or(0.0);
@@ -27,7 +27,7 @@ pub fn heston(
   for i in 1..n {
     s[i] = s[i - 1] + mu * s[i - 1] * dt + s[i - 1] * v[i - 1].sqrt() * correlated_bms[0][i - 1];
 
-    let random: f64 = match use_sym.unwrap_or(false) {
+    let random: f32 = match use_sym.unwrap_or(false) {
       true => eta * (v[i - 1]).abs().sqrt() * correlated_bms[1][i - 1],
       false => eta * (v[i - 1]).max(0.0).sqrt() * correlated_bms[1][i - 1],
     };
