+++ src/utils.rs
@@ -1,8 +1,8 @@
 use std::error::Error;
 
 pub trait Generator: Sync + Send {
-  fn sample(&self) -> Vec<f64>;
-  fn sample_par(&self) -> Vec<Vec<f64>>;
+  fn sample(&self) -> Vec<f32>;
+  fn sample_par(&self) -> Vec<Vec<f32>>;
 }
 
 pub trait Exporter: Sync + Send {
